Organized more commands into cogs
