+++ .gitignore
@@ -3,5 +3,7 @@ keys.py
 to_do.txt
 need_list.txt
 images
-restart_status.txt
 images/
+restart_status.txt
+afk
+afk/
