+++ jeeves.py
@@ -1,20 +1,18 @@
 from keys import *
 import discord
 import asyncio
-from discord.ext import commands
 import discord.utils
 import datetime
 import requests
-import urllib.request
-from urllib.parse import urlparse
-from os.path import splitext
+import aiohttp
+import os, sys
+
+from discord.ext import commands
 from lxml import html
 from lxml import etree
-import os, sys
-import aiohttp
 from urllib.parse import parse_qs
-import forecastio
 
+import forecastio
 import steamsearch
 
 api_key = weather_key
@@ -25,7 +23,7 @@ forecast = forecastio.load_forecast(api_key, lat, lng)
 
 steamsearch.set_key(steam_key, "anotherSession", cache=True)
 
-cogs = ['cogs.edward', 'cogs.weather', 'cogs.misc']
+cogs = ['cogs.edward', 'cogs.weather', 'cogs.misc', 'cogs.user_based', 'cogs.apartment', 'cogs.recipe']
 
 description = 'Jeeves Bot. Your personal helper bot.'
 
@@ -42,14 +40,19 @@ if restart_status == "0":
         print(bot.user.id)
         print('------')
         channel = bot.get_channel('276237909378465794')
-        await bot.send_message(channel, 'Jeeves Online.')
+        errors = []
         for cog in cogs:
             try:
                 bot.load_extension(cog)
             except (AttributeError, ImportError) as e:
                 print("```py\n{}: {}\n```".format(type(e).__name__, str(e)))
-                return
+                errors.append("true")
             print("{} loaded.".format(cog))
+
+        if "true" in errors:
+            await bot.send_message(channel, 'Jeeves Online. Some Cogs Misfunctioned on Start-Up.')
+        else:
+            await bot.send_message(channel, 'Jeeves Online. All Cogs Loaded.')
 
 elif restart_status == "1":
     bot = commands.Bot(command_prefix='!', self_bot=True)
@@ -61,14 +64,20 @@ elif restart_status == "1":
         print(bot.user.id)
         print('------')
         channel = bot.get_channel('276237909378465794')
-        await bot.send_message(channel, 'Self Bot Online.')
+        errors = []
         for cog in cogs:
             try:
                 bot.load_extension(cog)
             except (AttributeError, ImportError) as e:
                 print("```py\n{}: {}\n```".format(type(e).__name__, str(e)))
-            return
-        print("{} loaded.".format(cog))
+                errors.append("true")
+            print("{} loaded.".format(cog))
+
+        if "true" in errors:
+            await bot.send_message(channel, 'Self Bot Online. Some Cogs Misfunctioned on Start-Up.')
+        else:
+            await bot.send_message(channel, 'Self Bot Online. All Cogs Loaded.')
+
 else:
     bot = commands.Bot(command_prefix='!', description=description)
 
@@ -79,293 +88,19 @@ else:
         print(bot.user.id)
         print('------')
         channel = bot.get_channel('276237909378465794')
-        await bot.send_message(channel, 'Jeeves Online.')
+        errors = []
         for cog in cogs:
             try:
                 bot.load_extension(cog)
             except (AttributeError, ImportError) as e:
                 print("```py\n{}: {}\n```".format(type(e).__name__, str(e)))
-                return
+                errors.append("true")
             print("{} loaded.".format(cog))
 
-@bot.command(pass_context=True)
-async def wink(ctx):
-    channel = bot.get_channel(ctx.message.channel.id)
-    await bot.send_file(channel, "images/wink.png")
-    await bot.delete_message(ctx.message)
-
-
-@bot.command(pass_context=True)
-async def afk(ctx):
-    # 1 == afk | 0 == not afk
-    initial_author = ctx.message.author.name
-    initial_path = "afk/"+initial_author+".txt"
-    check_path = os.path.exists(initial_path)
-    if check_path == False:
-        file = open(initial_path, "w+")
-        file.write("0")
-        file.close()
-
-    initial_file = open(initial_path, 'r')
-    initial_status = initial_file.readline()
-    
-    if initial_status == "1":
-        await bot.delete_message(ctx.message)
-    elif initial_status == "0":
-        await bot.delete_message(ctx.message)
-        author =  ctx.message.author.name
-        path = "afk/"+author+".txt"
-        file = open(path, 'w')
-        file.write("1")
-        file.close()
-        await bot.say("*"+author+" is now AFK.*")
-    else:
-        await bot.delete_message(ctx.message)
-        author =  ctx.message.author.name
-        path = "afk/"+author+".txt"
-        file = open(path, 'w')
-        file.write("1")
-        file.close()
-        await bot.say("*"+author+" is now AFK.*")
-    
-
-@bot.command(pass_context=True)
-async def back(ctx):
-    initial_author = ctx.message.author.name
-    initial_path = "afk/"+initial_author+".txt"
-    initial_file = open(initial_path, 'r')
-    initial_status = initial_file.readline()
-
-    if initial_status == "1":
-        await bot.delete_message(ctx.message)
-        author = ctx.message.author.name
-        path = "afk/"+author+".txt"
-        file = open(path, 'w')
-        file.write("0")
-        file.close()
-        await bot.say("*"+author+" is no longer AFK.*")
-    elif initial_status == "0":
-        await bot.delete_message(ctx.message)
-    else:
-        await bot.delete_message(ctx.message)
-
-
-
-# --- SHOPPING LIST ---
-@bot.command()
-async def add_item(item):
-    need_file = open("need_list.txt", "r")
-    need_lines = need_file.readlines()
-    need_file.close()
-
-    listLen = len(need_lines)
-    need_lines.insert(listLen, item+"\n")
-    itemMessage = item + " has been added to your shopping list."
-    await bot.say(itemMessage)
-
-    lenList = len(need_lines)
-    file = open("need_list.txt", "w")
-    counter = 0
-    for _ in range(lenList):
-        line = need_lines[counter]
-        file.write(line)
-        counter += 1
-    file.close()
-
-@bot.command()
-async def remove_item(item):
-    need_file = open("need_list.txt", "r")
-    need_lines = need_file.readlines()
-    need_file.close()
-
-    if item in need_lines: need_lines.remove(item)
-    
-    file = open("need_list.txt", "r")
-    lines = file.readlines()
-    file.close()
-
-    file = open("need_list.txt", "w")
-    for line in lines:
-        if line!=item+"\n":
-            file.write(line)
-    file.close()
-    item_message = item + " has been removed from your shopping list."
-
-@bot.command()
-async def need_list():
-    need_list = "```Here is your current shopping list:\n\n"
-
-    file = open("need_list.txt", "r")
-    lines = file.readlines()
-
-    counter = 0
-    for _ in range(len(lines)):
-        need_list+="%s\n" % (lines[counter])
-        counter += 1
-    need_list+="```"
-    await bot.say(need_list)
-
-# --- TO DO LIST ---
-@bot.command()
-async def add_task(item):
-    need_file = open("to_do.txt", "r")
-    need_lines = need_file.readlines()
-    need_file.close()
-
-    listLen = len(need_lines)
-    need_lines.insert(listLen, item+"\n")
-    itemMessage = item + " has been added to your to do list."
-    await bot.say(itemMessage)
-
-    lenList = len(need_lines)
-    file = open("to_do.txt", "w")
-    counter = 0
-    for _ in range(lenList):
-        line = need_lines[counter]
-        file.write(line)
-        counter += 1
-    file.close()
-
-@bot.command()
-async def remove_task(item):
-    need_file = open("to_do.txt", "r")
-    need_lines = need_file.readlines()
-    need_file.close()
-
-    if item in need_lines: need_lines.remove(item)
-    
-    file = open("to_do.txt", "r")
-    lines = file.readlines()
-    file.close()
-
-    file = open("to_do.txt", "w")
-    for line in lines:
-        if line!=item+"\n":
-            file.write(line)
-    file.close()
-    task_message = item + " has been removed from the to do list."
-    await bot.say(task_message)
-
-@bot.command()
-async def to_do():
-    to_do = "```Here is your current to do list:\n\n"
-
-    file = open("to_do.txt", "r")
-    lines = file.readlines()
-
-    counter = 0
-    for _ in range(len(lines)):
-        to_do+="%s\n" % (lines[counter])
-        counter += 1
-    to_do+="```"
-    await bot.say(to_do)
-
-# --- RECIPES ---
-@bot.command(pass_context=True)
-async def recipe(ctx, *, ingredient):
-    url = "https://api.edamam.com/"
-    search = ingredient
-    search_addon = "search?q="+search
-    id_addon = "&app_id="+food_id
-    key_addon = "&app_key="+food_key
-    full_search = url+search_addon+id_addon+key_addon
-    channel = bot.get_channel(ctx.message.channel.id)
-    
-
-    r = requests.get(full_search)
-    recipe = r.json()
-    
-    url = recipe["hits"][0]["recipe"]["url"]
-    label = recipe["hits"][0]["recipe"]["label"]
-    image = recipe["hits"][0]["recipe"]["image"]
-    ingredients = recipe["hits"][0]["recipe"]["ingredientLines"]
-    calories = recipe["hits"][0]["recipe"]["calories"]
-    int(calories)
-
-    filename = image.split('/')[-1]
-    urllib.request.urlretrieve(image, "images/"+filename)
-    image_path = "images/"+filename
-
-    ingredient_list = []
-    for ing in ingredients:
-        ingredient_list.append(ing)
-
-    final_ingredients = ""
-    for ing in ingredient_list:
-        final_ingredients += ing+"\n"
-
-    recipe_message = """
-**Here is the most relevant recipe based on your search.** \n
-%s
-Calories: %d \n
-URL to recipe: <%s> \n
-    """ % (label, calories, url)
-
-    # await bot.say(recipe_message)
-    await bot.send_file(channel, image_path, content=recipe_message)
-
-@bot.command()
-async def get_recipes(ingredient, to_amount):
-    iteration = int(to_amount)
-
-    url = "https://api.edamam.com/"
-    search = ingredient
-    search_addon = "search?q="+search
-    id_addon = "&app_id="+food_id
-    key_addon = "&app_key="+food_key
-    from_addon = "&from=0"
-    to_addon = "&to="+str(to_amount)
-    full_search = url+search_addon+id_addon+key_addon+from_addon+to_addon
-
-    r = requests.get(full_search)
-    recipe = r.json()
-
-    url = recipe["hits"][1]["recipe"]["url"]
-    urls = []
-    counter = 0
-    for i in range(iteration):
-        urls.append(recipe["hits"][counter]["recipe"]["url"])
-        print(urls)
-        counter += 1
-
-    labels = []
-    labels_counter = 0
-    for i in range(iteration):
-        labels.append(recipe["hits"][labels_counter]["recipe"]["label"])
-        labels_counter += 1
-
-    recipe_counter = 0
-    recipes_message = "**Here are the most relevant recipes based on your search.**\n \n"
-    for i in range(iteration):
-        recipes_message += "%s \n<%s> \n \n" % (labels[recipe_counter], urls[recipe_counter])
-        recipe_counter += 1
-
-    await bot.say(recipes_message)
-
-@bot.command()
-async def ingredients(*,recipe):
-    url = "https://api.edamam.com/"
-    search = recipe
-    search_addon = "search?q="+search
-    id_addon = "&app_id="+food_id
-    key_addon = "&app_key="+food_key
-    full_search = url+search_addon+id_addon+key_addon
-
-    r = requests.get(full_search)
-    recipe = r.json()
-
-    url = recipe["hits"][1]["recipe"]["url"]
-    label = recipe["hits"][0]["recipe"]["label"]
-    ingredients = recipe["hits"][0]["recipe"]["ingredientLines"]
-
-    counter = 0
-    ingredients_message = "**%s**\n \n" % (label)
-    for i in range(len(ingredients)):
-        ingredients_message += recipe["hits"][0]["recipe"]["ingredientLines"][counter]+"\n"
-        counter += 1
-    ingredients_message += "\n<%s>" % (url)
-
-    await bot.say(ingredients_message)
+        if "true" in errors:
+            await bot.send_message(channel, 'Jeeves Online. Some Cogs Misfunctioned on Start-Up.')
+        else:
+            await bot.send_message(channel, 'Jeeves Online. All Cogs Loaded.')
     
 
 @bot.command()
@@ -772,14 +507,14 @@ async def my_background_task():
 
     rainChance = []
     for current in byDay.data:
-        rainChance.append(current.precipProbability)
+        rainChance.append(current.precipProbability*100)
 
         updateMsg = """```Here is your daily update.\n\n
 -----Weather-----\n
 %s\n
 Min Temperature: %s\n
 Max Temperature: %s\n
-Rain Chance: %d\n\n
+Rain Chance: %d%%\n\n
 
 -----Shopping List-----\n
 test
@@ -787,10 +522,10 @@ test
         ```
         """ % (weatherData[0], minTemp[0], maxTemp[0], rainChance[0])
 
-        while not bot.is_closed:
-            counter += 1
-            await bot.send_message(channel, updateMsg)
-            await asyncio.sleep(86400) # task runs once per day
+    while not bot.is_closed:
+        counter += 1
+        await bot.send_message(channel, updateMsg)
+        await asyncio.sleep(86400) # task runs once per day
 
 bot.loop.create_task(my_background_task())
